# openapi.partial.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/rag/retrieve": {
      "post": {
        "operationId": "RagController_retrieve",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rag"
        ]
      }
    },
    "/v1/rag/answer": {
      "post": {
        "operationId": "RagController_answer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rag"
        ]
      }
    },
    "/v1/rag/feedback": {
      "post": {
        "operationId": "RagController_feedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRagFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rag"
        ]
      }
    },
    "/v1/knowledge/sources": {
      "get": {
        "operationId": "KnowledgeController_listSources",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      },
      "post": {
        "operationId": "KnowledgeController_createSource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKnowledgeSourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/v1/knowledge/sources/{sourceId}": {
      "patch": {
        "operationId": "KnowledgeController_updateSource",
        "parameters": [
          {
            "name": "sourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKnowledgeSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
...
    "/v1/knowledge/documents/upload": {
      "post": {
        "operationId": "KnowledgeController_uploadDocument",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/v1/knowledge/documents/ingest-url": {
      "post": {
        "operationId": "KnowledgeController_ingestUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestDocumentUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/v1/knowledge/documents/{documentId}": {
      "get": {
        "operationId": "KnowledgeController_getDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/v1/knowledge/documents/{documentId}/chunks": {
      "get": {
        "operationId": "KnowledgeController_getDocumentChunks",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/v1/knowledge/documents/{documentId}/reprocess": {
      "post": {
        "operationId": "KnowledgeController_reprocessDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/v1/knowledge/documents/{documentId}/review": {
      "post": {
        "operationId": "KnowledgeController_reviewDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewKnowledgeDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
...
  },
  "info": {
    "title": "SaludDeUna API",
    "description": "Partial OpenAPI export for frontend type generation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RetrieveDto": {
        "type": "object",
        "properties": {}
      },
      "AnswerDto": {
        "type": "object",
        "properties": {}
      },
      "CreateRagFeedbackDto": {
        "type": "object",
        "properties": {}
      },
      "CreateKnowledgeSourceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateKnowledgeSourceDto": {
        "type": "object",
        "properties": {}
      },
      "IngestDocumentDto": {
        "type": "object",
        "properties": {}
      },
      "IngestDocumentUrlDto": {
        "type": "object",
        "properties": {}
      },
      "ReviewKnowledgeDocumentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}